동적 리뷰게시판 구현은 다 완료했으나 타임리프 또는 URL전체가 다 꼬여서 수정을 해야함 기능구현 실현을 동적 리뷰게시판을 먼저해야했음 다른 기능들이 꼬여서 URL 또는 타임리프 수정

+++ src/main/java/com/movie_board/movie_review/dto/PageDto.java
@@ -22,6 +22,7 @@ public class PageDto{
     private String filter; // 필터링 조건
 
 
+
     public PageDto(int total, int currentPage, int pageSize, String searchKeyword, String filter) {
         this.total = total;
         this.currentPage = currentPage;
@@ -31,7 +32,7 @@ public class PageDto{
 
 
         // 페이지 끝 번호 계산
-        this.pageEnd = (int) (Math.ceil(currentPage/ 5.0)) * 5;
+        this.pageEnd = (int) (Math.ceil((double) currentPage/ 5.0)) * 5;
 
         // 페이지 시작 시 번호 계산
         this.pageStart = this.pageEnd - 4;
